fix all problems pointed out
