+++ yamanaka/nextjs-app-dir/app/_common/components/index.tsx
@@ -1,0 +1,2 @@
+"use client";
+export * from "@chakra-ui/react";
