+++ yamanaka/nextjs-app-dir/app/_common/types.tsx
