+++ yamanaka/nextjs-app-dir/app/_components/ArticleList.tsx
@@ -1,6 +1,6 @@
-import { VStack } from "@/common/components";
+import { VStack } from "@/app/_common/components";
 import ArticleCard from "./ArticleCard";
-import { Article } from "@/common/types";
+import { Article } from "@/app/_common/types";
 
 export default function ArticleList({ articles }: { articles: Article[] }) {
   return (
