+++ yamanaka/nextjs-app-dir/app/_components/Footer.tsx
@@ -1,4 +1,4 @@
-import { Container, Box, Text } from "@/common/components";
+import { Container, Box, Text } from "@/app/_common/components";
 
 export default function Footer() {
   return (
