+++ yamanaka/nextjs-app-dir/app/_components/Header.tsx
@@ -1,4 +1,4 @@
-import { Box, Flex, Heading, Button } from "@/common/components";
+import { Box, Flex, Heading, Button } from "@/app/_common/components";
 import NextLink from "next/link";
 
 export default function Header() {
