--- yamanaka/nextjs-app-dir/app/articles/[slug]/components/index.tsx
@@ -1,3 +1,0 @@
-export {default as ArticleContent} from "./ArticleContent";
-export {default as Comments} from "./Comments";
-export {default as LoadingComments} from "./LoadingComments";
