--- yamanaka/nextjs-app-dir/app/common/components/index.tsx
@@ -1,2 +1,0 @@
-"use client";
-export * from "@chakra-ui/react";
