--- yamanaka/nextjs-app-dir/app/components/index.tsx
@@ -1,6 +1,0 @@
-export {default as ArticleCard} from "./ArticleCard";
-export {default as ArticleList} from "./ArticleList";
-export {default as Footer} from "./Footer";
-export {default as Header} from "./Header";
-export {default as Main} from "./Main";
-export {default as Provider} from "./Provider";
