+++ yamanaka/nextjs-app-dir/tsconfig.json
@@ -19,11 +19,11 @@
         "name": "next"
       }
     ],
-    "baseUrl":".",
-    "paths":{
-      "@/*":["app/*"],
+    "baseUrl": ".",
+    "paths": {
+      "@/*": ["./*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
-  "exclude": ["node_modules"],
+  "exclude": ["node_modules"]
 }
